Create README.md
Unacademy
A code to cut the background from the video and present it in new window with the filtered video. new win

How to use it
Clone the repository in your local system and run the video.py. You will get three windows where for your actual video, filtered video and a Trackbar will be shown. Set the trackbar to get the final video.